added c1x build directory, added MSVC Win64 project support

+++ src/share/vm/c1/c1_LIRGenerator.cpp
@@ -2427,7 +2427,7 @@ void LIRGenerator::do_Invoke(Invoke* x) {
 
       // Load CallSite object from constant pool cache.
       __ oop2reg(cpcache->constant_encoding(), tmp);
-      __ load(new LIR_Address(tmp, call_site_offset, T_OBJECT), tmp);
+      __ load(new LIR_Address(tmp, (int)call_site_offset, T_OBJECT), tmp);
 
       // Load target MethodHandle from CallSite object.
       __ load(new LIR_Address(tmp, java_dyn_CallSite::target_offset_in_bytes(), T_OBJECT), receiver);
